refactor: add bit_field.rs
