+++ crates/interpreter/src/lib.rs
@@ -148,6 +148,7 @@ mod parser;
 mod syntax;
 mod toctou;
 mod valid;
+mod bit_field;
 
 pub use error::{Error, Unsupported};
 pub use exec::{Call, InstId, RunAnswer, RunResult, Store, StoreId, Val, MEMORY_ALIGN};
